Autocut/requirements.txt
fastapi
moviepy
numpy
Pillow
pydantic
spaces
uvicorn
whisper